error handing if a user starts a focus session when there is a focus session happening or ending a focus session when it's already over

+++ functions/focus.py
@@ -21,34 +21,52 @@ email_reciver = EMAIL_RECIVER
     
 websites = ['www.youtube.com', "https://www.youtube.com", "www.instagram.com", "https://www.instagram.com/", "www.tiktok.com", "https://tiktok.com"]
 start = time.time()
-def start_focus():
-    os.system("taskkill /im chrome.exe /f")
-    with open(host_path, "r+") as fileptr: 
-        content = fileptr.read()
-        for website in websites: 
-            if website in content: 
-                pass
-            else: 
-                fileptr.write(redirect+"      "+website+"\n")            
-    webbrowser.open("https://www.notion.so/Assignments-Exams-0c3135345c6d4b6aabfc8a7150b55767")
-async def end_focus(user): 
-    end = time.time()
-    os.system("taskkill /im chrome.exe /f")
-    with open(host_path, 'r+') as fileptr:
-        content = fileptr.readlines()
-        fileptr.seek(0)
-        for line in content:
-            if not any(website in line for website in websites):
-                fileptr.write(line)
-  
-        # removing hostnames from the host file
-        fileptr.truncate()
-    duration = round(end - start)
-    await user.send(f"the focus session lasted {format_timespan(duration)}")
-    
-    subject = "Focus Session Completed 🚀"
-    body = f"I'm excited to share that Shashank has successfully completed a productive focus session. The session lasted {format_timespan(end-start)}. It's my part of keeping him focused, please respond with 'Yes' or 'No' to approve the end of his session."
 
+in_session = False
+
+async def start_focus(user):
+    global in_session
+    if in_session == True: 
+        await user.send("focus session is already started")
+
+    else:
+        in_session = True 
+        os.system("taskkill /im chrome.exe /f")
+        with open(host_path, "r+") as fileptr: 
+            content = fileptr.read()
+            for website in websites: 
+                if website in content: 
+                    pass
+                else: 
+                    fileptr.write(redirect+"      "+website+"\n")            
+        webbrowser.open("https://www.notion.so/Assignments-Exams-0c3135345c6d4b6aabfc8a7150b55767")
+        await user.send("blocked all entertainment websites")
+
+async def end_focus(user): 
+    global in_session
+    
+    if in_session == True:
+        in_session = False
+        end = time.time()
+        os.system("taskkill /im chrome.exe /f")
+        with open(host_path, 'r+') as fileptr:
+            content = fileptr.readlines()
+            fileptr.seek(0)
+            for line in content:
+                if not any(website in line for website in websites):
+                    fileptr.write(line)
+  
+                 # removing hostnames from the host file
+            fileptr.truncate()
+        duration = round(end - start)
+        await user.send(f"the focus session lasted {format_timespan(duration)}")
+         
+    else: 
+        await user.send("you haven't started a session")
+    
+    ## subject = "Focus Session Completed 🚀"
+    body = f"I'm excited to share that Shashank has successfully completed a productive focus session. The session lasted {format_timespan(end-start)}. It's my part of keeping him focused, please respond with 'Yes' or 'No' to approve the end of his session."
+    
     em = EmailMessage()
     em['From'] = email_sender
     em['To']= email_reciver
@@ -61,5 +79,6 @@ async def end_focus(user):
         smtp.login(email_sender, email_password)
         smtp.sendmail(email_sender, email_reciver, em.as_string())
 
-    await user.send("i have just sent an email to Mr. Ellareddy for the approval for the end of your focus session. You'll be notified as soon as i get a response.")
+        await user.send("i have just sent an email to Mr. Ellareddy for the approval for the end of your focus session. You'll be notified as soon as i get a response.")
+    
     webbrowser.open("https://www.notion.so/Assignments-Exams-0c3135345c6d4b6aabfc8a7150b55767")
